requirements.txt
PyPDF2
Pillow
numpy
pdf2image